+++ src/app/page.tsx
@@ -1,13 +1,20 @@
 import SearchBooksButton from '@/components/button/SearchBooksButton';
-import BookItems from '@/components/bookshelf/BookItems';
+import AuthenticatedBookItems from '@/components/bookshelf/AuthenticatedBookItems';
 import Cal from '@/components/Cal/Cal';
+import { auth } from '#auth';
+import TopPage from '@/components/top/TopPage';
 
 export default async function Home() {
-  return (
-    <div>
-      <SearchBooksButton />
-      <BookItems />
-      <Cal />
-    </div>
-  );
+  const session = await auth();
+  if (session?.user) {
+    return (
+      <div>
+        <SearchBooksButton />
+        <AuthenticatedBookItems />
+        <Cal />
+      </div>
+    );
+  } else {
+    return <TopPage />;
+  }
 }
