+++ src/components/bookshelf/BookItems.tsx
@@ -1,16 +1,5 @@
 import BookItem from './BookItem';
 import { SimpleGrid } from '@mantine/core';
-import { auth } from '@/auth';
-import axios from 'axios';
-
-type BookResponse = {
-  id: number;
-  title: string;
-  author: string;
-  cover_image_url: string;
-  created_at: string;
-  updated_at: string;
-};
 
 type Book = {
   id: number;
@@ -19,21 +8,11 @@ type Book = {
   coverImageUrl: string;
 };
 
-const BookItems = async () => {
-  const getBooks = async () => {
-    const session = await auth();
-    const params = { uid: session?.user?.id };
-    const res = await axios.get('http://localhost:3001/api/books', { params });
-    return res.data.map((book: BookResponse) => ({
-      id: book.id,
-      title: book.title,
-      author: book.author,
-      coverImageUrl: book.cover_image_url,
-    }));
-  };
+type BookItemsProps = {
+  bookItems: Book[]
+}
 
-  const bookItems = await getBooks();
-
+const BookItems = ({ bookItems }: BookItemsProps)=> {
   return (
     <div>
       {bookItems.length > 0 ? (
