+++ src/components/top/TopPage.tsx
@@ -1,0 +1,5 @@
+import SignInButton from "../button/SignInButton";
+
+export default function TopPage(){
+  return <SignInButton />
+}
